string/requirements.txt
numpy==2.1.2
scipy==1.14.1
matplotlib==3.9.2
scikit-image==0.24.0
Pillow==11.0.0
awslambdaric
boto3